Move shared components for adding accounts
