+++ packages/mobile-app/app/(drawer)/account/account-settings/account-select/index.tsx
@@ -3,7 +3,8 @@ import { StyleSheet } from "react-native";
 import { Stack, useRouter } from "expo-router";
 import { useFacade } from "@/data/facades";
 import { CurrencyUtils } from "@ironfish/sdk";
-import { Layout, Text, Button, Card } from "@ui-kitten/components";
+import { Layout, Text, Button, Card, Spinner } from "@ui-kitten/components";
+import React from "react";
 
 export default function AccountSelect() {
   const router = useRouter();
@@ -15,6 +16,12 @@ export default function AccountSelect() {
 
   const setActiveAccount = facade.setActiveAccount.useMutation();
 
+  const renderLoadingSkeleton = () => (
+    <Layout style={[styles.content, styles.loadingContainer]}>
+      <Spinner size="large" />
+    </Layout>
+  );
+
   return (
     <>
       <Stack.Screen
@@ -25,33 +32,35 @@ export default function AccountSelect() {
       />
       <Layout style={styles.container}>
         <Layout style={styles.content}>
-          {getAccountsResult.data?.map((account) => (
-            <Card
-              key={account.name}
-              style={styles.accountCard}
-              onPress={async () => {
-                const result = await setActiveAccount.mutateAsync({
-                  name: account.name,
-                });
-                console.log(`setActiveAccount: ${result}`);
-                router.dismissAll();
-              }}
-            >
-              <Layout style={styles.accountInfo}>
-                <Layout>
-                  <Text category="s1">{account.name}</Text>
-                  <Text category="p2" appearance="hint">
-                    {`${CurrencyUtils.render(account.balances.iron.confirmed)} $IRON`}
-                  </Text>
-                </Layout>
-                {account.active && (
-                  <Text status="success" category="c1">
-                    Active
-                  </Text>
-                )}
-              </Layout>
-            </Card>
-          ))}
+          {getAccountsResult.isLoading
+            ? renderLoadingSkeleton()
+            : getAccountsResult.data?.map((account) => (
+                <Card
+                  key={account.name}
+                  style={styles.accountCard}
+                  onPress={async () => {
+                    const result = await setActiveAccount.mutateAsync({
+                      name: account.name,
+                    });
+                    console.log(`setActiveAccount: ${result}`);
+                    router.dismissAll();
+                  }}
+                >
+                  <Layout style={styles.accountInfo}>
+                    <Layout>
+                      <Text category="s1">{account.name}</Text>
+                      <Text category="p2" appearance="hint">
+                        {`${CurrencyUtils.render(account.balances.iron.confirmed)} $IRON`}
+                      </Text>
+                    </Layout>
+                    {account.active && (
+                      <Text status="success" category="c1">
+                        Active
+                      </Text>
+                    )}
+                  </Layout>
+                </Card>
+              ))}
           <Button
             style={styles.addButton}
             onPress={() =>
@@ -87,5 +96,9 @@ const styles = StyleSheet.create({
   },
   addButton: {
     marginTop: 8,
+  },
+  loadingContainer: {
+    justifyContent: "center",
+    alignItems: "center",
   },
 });
